Introduce metrics.ServiceChecks
Move MarshalStrings from metricsserializer to metrics
